consolidate histogram code; minor styling fixes
